uint8

+++ src/main/java/com/ocrad/Pixmap.java
@@ -1,22 +1,22 @@
 package com.ocrad;
 
 public class Pixmap {
-	private byte data[];
+	private int data[];
 	private int height;
 	private int width;
 	private int mode;
 	
-	public Pixmap(byte[] data, int height, int width, int mode) {
+	public Pixmap(int[] data, int height, int width, int mode) {
 		super();
 		this.data = data;
 		this.height = height;
 		this.width = width;
 		this.mode = mode;
 	}
-	public byte[] getData() {
+	public int[] getData() {
 		return data;
 	}
-	public void setData(byte[] data) {
+	public void setData(int[] data) {
 		this.data = data;
 	}
 	public int getHeight() {
